utils: Add check for entrance repo path in arg ${1} used as docker volume

+++ utils/docker_entrance.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [[ -z ${1} ]]; then
+    echo "missing /path/to/entrance repo"
+    exit 1
+fi
+
 docker run \
 	--net=host \
 	--rm -it \
